moved existing source code fs extension to extra
