+++ lib/Extension/SourceCodeFilesystemExtra/Command/ClassSearchCommand.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Phpactor\Extension\SourceCodeFilesystem\Command;
+namespace Phpactor\Extension\SourceCodeFilesystemExtra\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
-use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilestem\Application\ClassSearch;
+use Phpactor\Extension\SourceCodeFilesystemExtra\SourceCodeFilestem\Application\ClassSearch;
 use Phpactor\Extension\Core\Console\Dumper\DumperRegistry;
-use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilesystemExtension;
+use Phpactor\Extension\SourceCodeFilesystemExtra\SourceCodeFilesystemExtraExtension;
 use Phpactor\Extension\Core\Console\Handler\FormatHandler;
 use Phpactor\Extension\Core\Console\Handler\FilesystemHandler;
 
@@ -38,7 +38,7 @@ class ClassSearchCommand extends Command
         $this->setDescription('Search for class by (short) name and return informations on candidates');
         $this->addArgument('name', InputArgument::REQUIRED, 'Source path or FQN');
         FormatHandler::configure($this);
-        FilesystemHandler::configure($this, SourceCodeFilesystemExtension::FILESYSTEM_COMPOSER);
+        FilesystemHandler::configure($this, SourceCodeFilesystemExtraExtension::FILESYSTEM_COMPOSER);
     }
 
     public function execute(InputInterface $input, OutputInterface $output)
