+++ tests/Unit/Extension/SourceCodeFilesystem/Rpc/ClassSearchHandlerTest.php
@@ -2,8 +2,8 @@
 
 namespace Phpactor\Tests\Unit\Extension\SourceCodeFilesystem\Rpc;
 
-use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilestem\Application\ClassSearch;
-use Phpactor\Extension\SourceCodeFilesystem\Rpc\ClassSearchHandler;
+use Phpactor\Extension\SourceCodeFilesystemExtra\SourceCodeFilestem\Application\ClassSearch;
+use Phpactor\Extension\SourceCodeFilesystemExtra\Rpc\ClassSearchHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\EchoResponse;
 use Phpactor\Extension\Rpc\Response\ReturnResponse;
